Repair!

+++ src/main/java/com/virtualelection/cloudproject1/repository/electionRepository.java
@@ -9,5 +9,5 @@ import org.springframework.stereotype.Repository;
 @Repository
 
 
-public interface electionRepository extends JpaRepository<Election, Long> {
+public interface electionRepository extends JpaRepository<Election, Integer> {
 }
